inertie robot + retour de force sur les codeurs + newton en v 2.29
	modified:   README
	modified:   src/control/control.c
	modified:   src/simulation/ArmTim.cxx
	modified:   src/simulation/ArmTim.h
	modified:   src/simulation/Motor.cxx
	modified:   src/simulation/Motor.h
	modified:   src/simulation/Pion.cxx
	modified:   src/simulation/Robot.cxx
	modified:   src/simulation/Robot.h
	modified:   src/simulation/Table.cxx
	modified:   src/test/strategy_test_control.c

+++ src/simulation/ArmTim.h
@@ -13,7 +13,7 @@ public:
 	ArmTim();
 	~ArmTim();
 
-	int32_t getPwm(int num);
+	float getPwm(int num);
 	void setEncoder(uint16_t val);
 };
